refactor: move intake/search domain logic out of app/utils into app/domain
