+++ backend/app/api/v1/endpoints/intake_sessions/answers/llm.py
@@ -7,6 +7,8 @@ from uuid import UUID
 from fastapi import APIRouter
 
 from app.core.deps import SupabaseSdkDep
+from app.domain.intake_criteria import normalize_merged_criteria
+from app.domain.intake_validation import compute_current_index
 from app.llm import (
     parse_user_input,
     resolve_next_intake_question,
@@ -21,8 +23,6 @@ from app.schemas.intake_sessions import (
     SubmitLlmIntakeInputRequest,
     SubmitLlmIntakeInputResponse,
 )
-from app.utils.intake_criteria import normalize_merged_criteria
-from app.utils.intake_validation import compute_current_index
 
 router = APIRouter()
