+++ backend/app/domain/__init__.py
@@ -1,0 +1,1 @@
+"""Domain logic: intake criteria/search rules that aren't generic utilities."""
